requirements.txt
django
gunicorn
django-heroku
scikit-learn
psycopg2-binary
dj-database-url
djangorestframework
